LPD-20678 Move package, ai-creator-openai-web
